packages/travel_agent/pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dharshan S Kotian", email = "dcodecraft@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["uv_build>=0.10.12,<0.11.0"]
build-backend = "uv_build"
